pyproject.toml
[project]
name = "topic-5-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi-cache2[redis]>=0.1.8",
    "fastapi[all]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "sqlalchemy>=2.0.40",
]
